Command line interface

A simple command line interface is generated with Comonicon.jl.

Index

Wannier.COMMAND_ENTRY_DOC_STUB

Wannier.band

Wannier.dis

Wannier.fermisurf

Wannier.maxloc

Wannier.optrot

Wannier.ptg

Wannier.splitvc

Wannier.trunc

Commands

Wannier.COMMAND_ENTRY_DOC_STUB — Constant

Julia package for Wannier functions.

Wannier.maxloc — Function

Maximally localize an isolated group of bands.

Args

seedname: seedname for win/amn/mmn/eig files

Options

-o, --output=<str>: filename for output amn. Default is seedname.maxloc.amn

-m, --maxiter=<int>: max number of iterations. Default is 50

Wannier.dis — Function

Maximally localize a group of entangled bands.

Args

seedname: seedname for win/amn/mmn/eig files

Options

-o, --output=<str>: filename for output amn. Default is seedname.dis.amn

-m, --maxiter=<int>: max number of iterations. Default is 50

Wannier.ptg — Function

Generate parallel transport gauge Wannier functions.

Args

seedname: seedname for win/amn/mmn/eig files

Options

-o, --output: filename for output amn. Default is seedname.ptg.amn

Wannier.splitvc — Function

Split valence and conduction Wannier functions.

Usually start from a Wannierization of valence+conduction bands. Then this command split WFs into two independent groups.

Args

seedname: seedname for win/amn/mmn/eig files

Options

--nval: number of valence WFs. Default is n_wann ÷ 2

--outdir-val: dirname for output valence amn/mmn/eig. Default is val

--outdir-cond: dirname for output conduction amn/mmn/eig. Default is cond

--config: config file for splitvc command, e.g. toml [groups] indices = [ [ 1, 2,], [ 3, 4, 5, 6,], [ 7, 8,], ] outdirs = [ "val_1", "val_2", "cond_3",]

Flags

--run-disentangle: read amn and run disentangle first, otherwise read chk to get unitary matrices from n_bands to n_wann

--run-optrot: max localize w.r.t. single unitary matrix after parallel transport. Should further reduce the spread and much closer to the true max localization.

--run-maxloc: run a final max localize w.r.t. all kpoints. Should reach the true max localization.

--rotate-unk: generate unk files for valence and conduction, for plotting WFs

--binary: write amn/mmn/eig/unk in Fortran binary format

Wannier.optrot — Function

Maximally localize w.r.t a single unitary rotation of all the kpoints.

Usually should start from parallel transport gauge amn, where the gauge are already smoothened w.r.t. kpoints. However, there is sitll a global unitary transformation freedom, which will be minimized by this optimal rotation function.

Args

seedname: seedname for win/amn/mmn/eig files

Options

-o, --output=<str>: filename for output amn. Default is seedname.optrot.amn

-m, --maxiter=<int>: max number of iterations. Default is 50

Wannier.trunc — Function

Truncate number of bands in mmn/eig/unk files.

Args

seedname: seedname for mmn/eig files

keep-bands: indices of bands to be kept, start from 1

Options

--outdir: dirname for output mmn/eig. Default is truncate

Flags

--rotate-unk: also truncate unk files, for plotting WFs

Wannier.band — Function

Interpolate band structure.

Args

seedname: seedname for amn/mmn/eig/chk files

Options

--out: output seedname for bands.dat. Default is wjl

--amn: amn filename. If not given, default is read chk.fmt file

Wannier.fermisurf — Function

Interpolate Fermi surface.

Args

seedname: seedname for amn/mmn/eig/chk files or tb/wsvec.dat file

Options

--nk: number of interpolated kpoints along each reciprocal lattice vector. Default is 50, on output bxsf, there will be 51 points (the last point is the periodic image of the first point).

--dk: spacing of interpolated kpoints along each reciprocal lattice vector, in unit of Å^-1. Default is a negative number, meaning I will use --nk instead; if dk > 0, dk will take precedence. The number of interpolated kpoints is the nearest integer to a/dk, where a is the norm of each reciprocal lattice vector. Also plus 1 due to the same reason in --nk.

--ef: Fermi energy. Default is 0.0.

--out: output seedname for bxsf. Default is wjl

--amn: amn filename. If not given, default is read chk.fmt file

Flags

--tb: read tb/wsvec.dat file instead of amn/mmn/eig/chk files. Default is false.