Wannierize

These are some Wannierization algorithms.

Contents

Wannierize

Contents

Index

Disentanglement

Maximal localization

Parallel transport

Optimal rotation

Splitting the Model

Co-optimization of spin-polarized WFs

Constraining WF centers

Index

Wannier.MagModel

Wannier.GU_to_GX_GY

Wannier.U_to_X_Y

Wannier.XY_to_X_Y

Wannier.X_Y_to_U

Wannier.X_Y_to_XY

Wannier.check_frozen_bands

Wannier.choose_pole

Wannier.compute_error

Wannier.disentangle

Wannier.disentangle

Wannier.disentangle_center

Wannier.get_fg!_center_disentangle

Wannier.get_fg!_disentangle

Wannier.get_fg!_disentangle

Wannier.get_fg!_maxloc

Wannier.get_fg!_rotate

Wannier.get_frozen_bands

Wannier.get_frozen_proj

Wannier.interpolate_vec

Wannier.matrix_parallel_transport

Wannier.matrix_transport

Wannier.max_localize

Wannier.omega_updn

Wannier.opt_rotate

Wannier.orthonorm_freeze

Wannier.overlap_updn

Wannier.overlap_updn_grad

Wannier.overlap_updn_grad

Wannier.parallel_transport

Wannier.propagate!

Wannier.rotate_U

Wannier.set_frozen_degen!

Wannier.set_frozen_proj!

Wannier.set_frozen_win!

Wannier.split_eig

Wannier.split_eig

Wannier.split_model

Wannier.split_model

Wannier.split_unk

Wannier.split_unk

Wannier.split_wannierize

Wannier.split_wannierize

Wannier.zero_froz_grad!

Disentanglement

Wannier.GU_to_GX_GY — Method

GU_to_GX_GY(G, X, Y, frozen)

Compute dΩ/dX and dΩ/dY from dΩ/dU.

Acutally they are the conjugate gradients, e.g., $\frac{d \Omega}{d U^*}$.

Arguments

G: n_bands * n_wann * n_kpts array for gradient dΩ/dU

X: n_wann * n_wann * n_kpts array for X

Y: n_bands * n_wann * n_kpts array for Y

frozen: n_bands * n_kpts BitMatrix for frozen bands

Wannier.U_to_X_Y — Method

U_to_X_Y(U::Array{T,3}, frozen::BitMatrix) where {T<:Complex}

Convert the U layout to the (X, Y) layout.

See also X_Y_to_U.

Arguments

U: n_bands * n_wann * n_kpts

frozen: n_bands * n_kpts

Wannier.XY_to_X_Y — Method

XY_to_X_Y(XY::Matrix{T}, n_bands::Int, n_wann::Int)

Convert the XY layout to the (X, Y) layout.

See also X_Y_to_U.

Arguments

XY: n_bands * n_wann * n_kpts contiguous array

n_bands: number of bands, to be used to reshape XY

n_wann: number of wannier functions, to be used to reshape XY

Wannier.X_Y_to_U — Method

X_Y_to_U(X::Array{T,3}, Y::Array{T,3})

Convert the (X, Y) layout to the U layout.

There are three formats: U, (X, Y), and XY stored contiguously in memory. For each kpoint,

U: size(U) = (n_bands, n_wann), the format used in the rest of the code

(X, Y): size(X) = (n_wann, n_wann), size(Y) = (n_bands, n_wann), intermediate format

XY: this is the format used in the optimizer

Wannier.X_Y_to_XY — Method

X_Y_to_XY(X::Array{T,3}, Y::Array{T,3}) where {T<:Complex}

Convert the (X, Y) layout to the XY layout.

See also X_Y_to_U.

Wannier.check_frozen_bands — Method

check_frozen_bands(frozen_bands, n_wann)

Sanity check that the number of frozen bands at each kpoint <= n_wann.

Arguments

frozen_bands: the BitMatrix of frozen bands

n_wann: the number of wannier functions

Wannier.disentangle — Method

disentangle(model; random_gauge=false, f_tol=1e-7, g_tol=1e-5, max_iter=200, history_size=3)

Run disentangle on the Model.

Arguments

model: model

Keyword arguments

random_gauge: use random U matrices as initial guess

f_tol: tolerance for spread convergence

g_tol: tolerance for gradient convergence

max_iter: maximum number of iterations

history_size: history size of LBFGS

Wannier.get_fg!_disentangle — Method

get_fg!_disentangle(model::Model)

Return a tuple of two functions (f, g!) for spread and gradient, respectively.

Wannier.get_frozen_bands — Function

get_frozen_bands(E, dis_froz_max, dis_froz_min)

Generate a BitMatrix of frozen bands by checking the two frozen windows.

Arguments

E: the energy eigenvalues of the Hamiltonian

dis_froz_max: the upper bound of the frozen window

dis_froz_min: the lower bound of the frozen window

Note

The dis_froz_max and dis_froz_min work similarly as Wannier90.

Wannier.get_frozen_proj — Method

get_frozen_proj(E, U, dis_proj_max)

Get frozen bands according to band projectability.

Arguments

E: the energy eigenvalues of the Hamiltonian

U: the gauge rotation matrices

dis_proj_max: the upper bound projectability. Bands with projectability >= dis_proj_max are frozen.

Note

The band projectability for band $n$ at kpoint $\bm{k}$ is calculated by $p_{n \bm{k}} = \sum_{m=1}^{m=n_{wann}} | U_{nm \bm{k}} |^2$, and usually each element $p_{n \bm{k}} \in [0.0, 1.0]$. In such cases, the dis_proj_max is usually set to sth. like 0.9 to freeze high-projectability bands.

Wannier.orthonorm_freeze — Method

orthonorm_freeze(U, frozen)

Normalize and freeze a block of a matrix.

Conditions:

Block form: U = vcat(Uf, Ur)

Semiunitary

U' * U = I

Uf' * Uf + Ur' * Ur = I

Frozen: Uf * Uf' = I

Also: Uf * Ur' = 0

Strategy:

orthogonalize Uf

project Uf out of Ur

orthogonalize the range of Ur

Arguments

U: the matrix to be orthonormalized and frozen

frozen: the BitVector specifying which bands are frozen

Wannier.set_frozen_degen! — Method

set_frozen_degen!(frozen_bands, E, atol=1e-4)

Freeze bands which are degenerate with the frozen_bands.

In some cases, we might want to freeze the whole set of degenerated eigen vectors.

Arguments

frozen_bands: the BitMatrix of frozen bands

E: the energy eigenvalues of the Hamiltonian

atol: the tolerance of degeneracy

Wannier.set_frozen_proj! — Method

set_frozen_proj!(model, dis_proj_max; degen=false, degen_atol=1e-4)

Set frozen bands of the Model according to projectability.

Arguments

model: the Model to be set

dis_proj_max: the upper bound projectability. Bands with projectability >= dis_proj_max are frozen.

Keyword Arguments

degen: whether to freeze the whole set of degenerated eigen vectors

degen_atol: the tolerance of degeneracy

Wannier.set_frozen_win! — Method

set_frozen_win!(model, dis_froz_max, dis_froz_min=-Inf; degen=false, degen_atol=1e-4)

Set frozen bands of the Model according to two energy windows.

Arguments

model: the Model to be set

dis_froz_max: the upper bound of the frozen window

dis_froz_min: the lower bound of the frozen window

Keyword Arguments

degen: whether to freeze the whole set of degenerated eigen vectors

degen_atol: the tolerance of degeneracy

Wannier.zero_froz_grad! — Method

zero_froz_grad!(G, frozen)

Set gradient of frozen bands to 0.

This is used in test.

Arguments

G: gradient of the spread, in XY layout

frozen: BitMatrix for frozen bands, n_bands * n_kpts

Maximal localization

Wannier.get_fg!_maxloc — Method

get_fg!_maxloc(model::Model)

Return a tuple of two functions (f, g!) for spread and gradient, respectively.

Wannier.max_localize — Method

max_localize(model; f_tol=1e-7, g_tol=1e-5, max_iter=200, history_size=3)

Maximally localize spread functional w.r.t. all kpoints on a unitary matrix manifold.

Arguments

model: model

Keyword arguments

f_tol: tolerance for spread convergence

g_tol: tolerance for gradient convergence

max_iter: maximum number of iterations

history_size: history size of LBFGS

Parallel transport

Wannier.compute_error — Method

compute_error(model, U::Array{Complex{T},3})

Compute the smoothness error of the gauge.

Wannier.parallel_transport — Method

parallel_transport(model::Model{T}; use_U=false, log_interp=false)

Parallel transport the gauge from the first kpoint to all other kpoints.

Assumptions:

the kpoints are contained in a N1 * N2 * N3 cartesian grid

the neighbor list must contain the six cartesian neighbors along x, y, z directions

Arguments

model: model

Keyword arguments

use_U: use the gauge U instead of random matrix

log_interp: use logarithmic interpolation method

Wannier.choose_pole — Method

Choose the column and the target point to contract the vector path of the columns of matrix_path

Wannier.interpolate_vec — Method

Create an interpolation path between x and y, unit vectors by normalizing the linear interpolation of parameter t[i].

Wannier.matrix_parallel_transport — Method

Compute the parallel transport of the matrix path (matrixpath) along the homotopy of the frame (framepath) that starts with the columns (columns[]) of matrix_path.

size(framepath) = (ncol, ncol, nk, nt) size(matrixpath) = (ncol, ncol, n_k)

Wannier.matrix_transport — Method

Contract the matrix_path to a single matrix point in the space of unitaries using parallel transport.

i.e., contract Obs(k) matrices to constant vectors.

size(matrixpath) = nwann x nwann x nk t: vector of kpoint indexes along a different k direction

Wannier.propagate! — Method

Propagate U, defined at the first kpt, to the given list of kpts. Those must be neighbors, and only the first kpoint is assumed to have been rotated.

Arguments

U: n_wann * n_wann * n_kpts gauge matrices

kpts: list of kpoints along which to propagate the gauge matrices

dk: the distance between two consecutive kpoints, in fractional coordinates

M: overlap matrices

kpoints: BVectors.kpoints, kpoint coordinates of the grid

kpb_k: BVectors.kpb_k

kpb_G: BVectors.kpb_G

Optimal rotation

Wannier.get_fg!_rotate — Method

get_fg!_rotate(model::Model)

Return a tuple of two functions (f, g!) for spread and gradient, respectively.

Wannier.opt_rotate — Method

opt_rotate(model; f_tol=1e-7, g_tol=1e-5, max_iter=200, history_size=3)

Maximally localize spread functional w.r.t. single unitary matrix W.

Arguments

model: model

Keyword arguments

f_tol: tolerance for spread convergence

g_tol: tolerance for gradient convergence

max_iter: maximum number of iterations

history_size: history size of LBFGS

Wannier.rotate_U — Method

rotate_U(U::Array{T,3}, W::Matrix{T}) where {T<:Complex}

Rotate the U matrices at each kpoint by the same W matrix.

$\forall \bm{k}$, $U_{\bm{k}} W$

Useful once we have the optimal rotation matrix W, then update the initial U matrices by rotating them by W.

Splitting the Model

Wannier.split_eig — Method

split_eig(E, U, eig_groups)

Split eigenvalues into several groups.

The separation is done by

construct Wannier gauge Hamiltonian, $H_{\bm{k}} = U_{\bm{k}}^\dagger [\epsilon_{n \bm{k}}] U_{\bm{k}}$

diagonalize the Hamiltonian, the eigenvalues are sorted in ascending order, and they are split into several groups according to the indices in eig_groups.

Arguments

E: eigenvalues

U: (semi-)Unitary matrices gauge transformation

eig_groups: a Vector, in which each element is a Vector of indices of eigenvalues that belong to the same group.

Tip

For example, if one wants to split valence+conduction into two groups, valence and conduction, respectively, then eig_groups should be [[1:n_val], [n_val+1:n_wann]]. Usually, the U are the maximally localized gauge matrices from a valence+conduction band Wannierization, this function split the gauge matrices U into two groups by diagonalizing the Hamiltonian, so we have two set of eigenvalues and gauge matrices for valence and conduction bands, respectively. However, the diagonalization introduce random gauges, so the returned two gauge matrices for valence and conduction are bad, we new to run a parallel transport to smoothen the gauges.

Wannier.split_eig — Method

split_eig(E, U, n_val)

Split eigenvalues into two groups.

The separation is done by

construct Wannier gauge Hamiltonian, $H_{\bm{k}} = U_{\bm{k}}^\dagger [\epsilon_{n \bm{k}}] U_{\bm{k}}$

diagonalize the Hamiltonian, the eigenvalues are sorted in ascending order, so that the first n_val eigenvalues are the occupied states, and the rest are the unoccupied states.

Arguments

E: eigenvalues

U: (semi-)Unitary matrices gauge transformation

n_val: number of valence states

Wannier.split_model — Method

split_model(model, n_val)

Split the Model into two Models.

Arguments

model: the Model to be split

n_val: number of valence WFs

Wannier.split_model — Method

split_model(model, eig_groups)

Split the Model into several Models.

Arguments

model: the Model to be split

eig_groups: a Vector, in which each element is a Vector of indices of eigenvalues that belong to the same group.

Note

Rotation eigenvectors are also returned, useful for further rotation of UNK files or other operators.

Wannier.split_unk — Method

split_unk(dir, Us, outdirs)

Rotate UNK files.

These are large matrices, so we read/write to disk for each kpoint sequentially, inside the function.

Arguments

dir: directory where UNK files are stored

Us: a Vector of Wannier (semi-)unitary gauge matrices for rotating band groups

outdirs: a Vector of output directories for each band group

Keyword arguments

binary: whether to write in Fortran binary format

Wannier.split_unk — Method

split_unk(dir, Uv, Uc, outdir_val="val", outdir_cond="cond")

Rotate UNK files.

These are large matrices, so we read/write to disk for each kpoint sequentially, inside the function.

Arguments

dir: directory where UNK files are stored

Uv: the Wannier (semi-)unitary matrix for rotating valence bands

Uc: the Wannier (semi-)unitary matrix for rotating conduction bands

outdir_val: output directory for valence bands

outdir_cond: output directory for conduction bands

Keyword arguments

binary: whether to write in Fortran binary format

Wannier.split_wannierize — Method

split_wannierize(model::Model, n_val::Int)

Split the model and run parallel transport to smoothen the gauge.

Arguments

model: the Model to be split

n_val: number of valence WFs

Wannier.split_wannierize — Method

split_wannierize(model::Model, eig_groups)

Split the model and run parallel transport to smoothen the gauge.

Arguments

model: the Model to be split

eig_groups: a Vector, in which each element is a Vector of indices of eigenvalues that belong to the same group.

Note

Return two separated Models and rotation matrices which are useful for UNK files or other operators.

Co-optimization of spin-polarized WFs

Wannier.MagModel — Type

Model for spin polarized system with constraint.

Traditionally, we run two independent Wannierizations for spin up and spin down. Here we add a constraint to maximally overlap the spin-up and spin-down WFs, so that they map one-by-one to each other.

Wannier.disentangle — Method

disentangle(model; f_tol=1e-7, g_tol=1e-5, max_iter=200, history_size=3)

Run disentangle on a MagModel.

Arguments

model: MagModel

λ: Lagrange multiplier of the ↑↓ overlap term

Keyword arguments

f_tol: tolerance for spread convergence

g_tol: tolerance for gradient convergence

max_iter: maximum number of iterations

history_size: history size of LBFGS

Wannier.get_fg!_disentangle — Function

get_fg!_disentangle(model::MagModel)

Return a tuple of two functions (f, g!) for spread and gradient, respectively.

Wannier.omega_updn — Method

omega_updn(M)

Compute QPPM Eq. 8.

Arguments

M: the overlap matrix between up and down WFs, size: (nwann, nwann), should be the matrix returned from overlap_updn.

Wannier.overlap_updn — Method

overlap_updn(up::Model{T}, dn::Model{T}, M::Array{Complex{T},3}) where {T<:Real}

Compute the overlap between up and down WFs.

Actually N - Ω↑↓, according to QPPM Eq. 8, where N = n_wann.

Arguments

M: the MagModel.M matrices, size (nbands, nbands, n_kpts)

Uup: the up gauge matrices, size: (nbands, nwann, n_kpts)

Udn: the down gauge matrices, size: (nbands, nwann, n_kpts)

Wannier.overlap_updn_grad — Method

overlap_updn_grad(model::MagModel, Uup, Udn)

Compute gradients of overlap_updn.

$\frac{d \Omega}{d U^{\uparrow}}$ and $\frac{d \Omega}{d U^{\downarrow}}$.

TODO: this is actually the gradient of Tr[overlap_updn]

Arguments

M: the MagModel.M matrices, size (nbands, nbands, n_kpts)

Uup: the up gauge matrices, size: (nbands, nwann, n_kpts)

Udn: the down gauge matrices, size: (nbands, nwann, n_kpts)

Wannier.overlap_updn_grad — Method

overlap_updn_grad(model::MagModel, Xup, Yup, Xdn, Ydn)

Compute gradient of overlap_updn.

$\frac{d \Omega}{d X^{\uparrow}}$, $\frac{d \Omega}{d Y^{\uparrow}}$, $\frac{d \Omega}{d X^{\downarrow}}$, $\frac{d \Omega}{d Y^{\downarrow}}$.

Constraining WF centers

Disentanglement

Maximal localization

Co-optimization of spin-polarized WFs

Wannier.disentangle_center — Method

disentangle(model; f_tol=1e-7, g_tol=1e-5, max_iter=200, history_size=3)

Run disentangle on a MagModel, with center constraints.

Arguments

model: MagModel

r₀: 3 * n_wann, WF centers in cartesian coordinates

λc: Lagrange multiplier of the center term

λs: Lagrange multiplier of the spin-up and spin-down overlap term

Keyword arguments

f_tol: tolerance for spread convergence

g_tol: tolerance for gradient convergence

max_iter: maximum number of iterations

history_size: history size of LBFGS

Wannier.get_fg!_center_disentangle — Method

get_fg!_center_disentangle(model::MagModel, r₀, λc, λs)

Return a tuple of two functions (f, g!) for spread and gradient, respectively.