Real space WFs

This page lists functions for processing WFs in real space.

Normally operators are computed in reciprocal space, but sometimes it might be useful to evaluate them in real space. For example, computing higher moment of WFs.

Contents

Real space WFs

Contents

Index

Real-space grid struct

Read/write real-space WFs

Evaluate operators in real space

Index

Wannier.RGrid

Wannier.RGrid

Wannier.cartesianize_xyz

Wannier.origin

Wannier.span_vectors

Real-space grid struct

Wannier.RGrid — Type

struct RGrid(basis, X, Y, Z)

Represents a regular grid of points.

Fields

basis: each column is a basis vector of the grid, usually just the lattice vectors.

X: nx * ny * nz array of fractional coordinate w.r.t basis, the x coordinate of each point in the grid.

Y: nx * ny * nz array of fractional coordinate w.r.t basis, the y coordinate of each point in the grid.

Z: nx * ny * nz array of fractional coordinate w.r.t basis, the z coordinate of each point in the grid.

Tip

The X, Y, and Z are usually generated by LazyGrids.ndgrid. They are fractional and can be outside of [0, 1), so the basis are not necessarily the spanning vectors of the grid. See also origin and span_vectors.

Usually the grid does not contain the periodically repeated points, e.g., the x coordinate can be [0.0, 0.25, 0.5, 0.75] without 1.0 which is repeatition of 0.0.

Wannier.RGrid — Method

RGrid(basis, origin, X, Y, Z)

Construct a regular grid of points.

Arguments

basis: each column is a basis vector of the grid, usually just the lattice vectors.

origin: 3, origin of the grid in cartesian coordinates

X: nx vector of fractional coordinate w.r.t basis, the x coordinate of each point in the grid.

Y: ny vector of fractional coordinate w.r.t basis, the y coordinate of each point in the grid.

Z: nz vector of fractional coordinate w.r.t basis, the z coordinate of each point in the grid.

Wannier.cartesianize_xyz — Method

cartesianize_xyz(rgrid::RGrid)

Return X, Y, Z in cartesian coordinates.

The size of the returned X, Y, and Z are nx * ny * nz.

Wannier.origin — Method

origin(rgrid::RGrid)

Get the origin in cartesian coordinates, i.e. the 1st point, of the RGrid.

Wannier.span_vectors — Method

span_vectors(rgrid::RGrid)

Get the spanning vectors of the RGrid.

Each column in the returned matrix is a spanning vector.

Note

There is no limit constraint on the fractional coordinates X, Y, and Z, so the spanning vectors are not necessarily the basis vectors, they can be fractional of the basis vectors, or multiple times of that.

Read/write real-space WFs

Evaluate operators in real space